requirements.txt
Django==2.1
djangorestframework==3.11.2
Faker==2.0.4
factory-boy==2.12.0
python-dotenv==0.17.0
celery==5.0.5
psycopg2-binary>=2.8
drf-yasg==1.20.0